requirements.txt
flask
flask-session
gitpython
markdown
openai
requests
secrets